requirements.txt
# Geode Six — Python Dependencies
# Install: pip install -r requirements.txt

# API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# File handling
python-multipart>=0.0.6
python-docx>=1.1.0
openpyxl>=3.1.2
PyPDF2>=3.0.0
Pillow>=10.0.0

# HTTP client (for Ollama API)
httpx>=0.25.0

# System monitoring
psutil>=5.9.0

# Vector database
chromadb>=0.4.0

# Embedding fallback (if nomic-embed-text unavailable via Ollama)
# sentence-transformers>=2.2.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
